requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
assemblyai==0.23.0
attrs==23.2.0
beanie==1.25.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
hyperframe==5.2.0
idna==2.10
lazy-model==0.2.0
motor==3.3.2
multidict==6.0.5
openai==0.28.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.1
starlette==0.36.3
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
